Add ProviderCard, PersonDetailsPanel, and RecognizedPersonListTile widgets; enable editable names for tracked faces and enhance UI with status indicators.
